+++ packages/@eventual/aws-cdk/src/events.ts
@@ -1,36 +1,15 @@
 import { ENV_NAMES } from "@eventual/aws-runtime";
 import { internal } from "@eventual/core";
-import { aws_eventschemas, aws_iam, Lazy, Resource } from "aws-cdk-lib";
-import { EventBus, IEventBus, Rule } from "aws-cdk-lib/aws-events";
-import { LambdaFunction } from "aws-cdk-lib/aws-events-targets";
-import {
-  CompositePrincipal,
-  IGrantable,
-  IPrincipal,
-} from "aws-cdk-lib/aws-iam";
-import { Function, FunctionProps } from "aws-cdk-lib/aws-lambda";
-import { IQueue, Queue } from "aws-cdk-lib/aws-sqs";
+import { aws_eventschemas, Lazy, Resource } from "aws-cdk-lib";
+import { EventBus, IEventBus } from "aws-cdk-lib/aws-events";
+import { IGrantable } from "aws-cdk-lib/aws-iam";
+import { Function } from "aws-cdk-lib/aws-lambda";
 import { Construct } from "constructs";
 import type { OpenAPIObject, SchemaObject } from "openapi3-ts";
 import type { BuildOutput } from "./build";
-import { IService } from "./service";
-import { IServiceApi } from "./service-api";
-import { ServiceFunction, ServiceFunctionProps } from "./service-function";
 import { grant } from "./grant";
-import { computeDurationSeconds } from "@eventual/core-runtime";
-import { Duration } from "aws-cdk-lib";
-import type { KeysOfType } from "./utils";
 
-export type EventHandlerNames<Service> = KeysOfType<
-  Service,
-  { kind: "EventHandler" }
->;
-
-export type SubscriptionProps<Service> = {
-  [eventHandler in EventHandlerNames<Service>]?: EventHandlerProps;
-};
-
-export interface EventsProps<Service = any> {
+export interface EventsProps {
   /**
    * The built service describing the event subscriptions within the Service.
    */
@@ -39,52 +18,17 @@ export interface EventsProps<Service = any> {
    * The name of the Service this {@link Events} repository belongs to.
    */
   readonly serviceName: string;
-  /**
-   * Optional environment variables to add to the {@link Events.defaultHandler}.
-   *
-   * @default - no extra environment variables
-   */
-  readonly environment?: Record<string, string>;
-  /**
-   * Configuration for individual Event Handlers created with `onEvent`.
-   */
-  readonly subscriptions?: SubscriptionProps<Service>;
-  readonly service: IService;
-  readonly api: IServiceApi;
 }
 
-export interface EventHandlerProps
-  extends Omit<Partial<FunctionProps>, "code" | "handler" | "functionName"> {}
-
-export class Events<Service> extends Construct implements IGrantable {
+export class Events<Service> extends Construct {
   /**
    * The {@link EventBus} containing all events flowing into and out of this {@link Service}.
    */
   public readonly bus: IEventBus;
 
-  /**
-   * Individual Event Handler Lambda Functions handling only events they subscribe to. These handlers
-   * are individually bundled and tree-shaken for optimal performance and may contain their own custom
-   * memory and timeout configuration.
-   */
-  public readonly subscriptions: {
-    [handler in EventHandlerNames<Service>]: Function;
-  };
-
-  public get handlers(): Function[] {
-    return Object.values(this.subscriptions);
-  }
-
-  /**
-   * A SQS Queue to collect events that failed to be handled.
-   */
-  public readonly deadLetterQueue: IQueue;
-
-  public readonly grantPrincipal: IPrincipal;
-
   private readonly serviceName: string;
 
-  constructor(scope: Construct, id: string, private props: EventsProps) {
+  constructor(scope: Construct, id: string, props: EventsProps) {
     super(scope, id);
 
     this.serviceName = props.serviceName;
@@ -92,89 +36,6 @@ export class Events<Service> extends Construct implements IGrantable {
     this.bus = new EventBus(this, "Bus", {
       eventBusName: props.serviceName,
     });
-
-    this.deadLetterQueue = new Queue(this, "DeadLetterQueue");
-
-    const functionProps: Partial<ServiceFunctionProps> = {
-      deadLetterQueueEnabled: true,
-      deadLetterQueue: this.deadLetterQueue,
-      retryAttempts: 2,
-      environment: props.environment,
-    };
-
-    const role = new aws_iam.Role(this, "DefaultSubscriptionRole", {
-      assumedBy: new aws_iam.ServicePrincipal("lambda.amazonaws.com"),
-      managedPolicies: [
-        aws_iam.ManagedPolicy.fromAwsManagedPolicyName(
-          "service-role/AWSLambdaBasicExecutionRole"
-        ),
-      ],
-    });
-
-    // create a Construct to safely nest bundled functions in their own namespace
-    const bundledHandlers = new Construct(this, "BundledHandlers");
-
-    this.subscriptions = Object.fromEntries(
-      Object.values(props.build.subscriptions).map((func) => {
-        const sub = func.spec;
-        const handler = new ServiceFunction(bundledHandlers, func.spec.name, {
-          code: props.build.getCode(func.file),
-          functionName: `${props.serviceName}-event-${sub.name}`,
-          ...functionProps,
-          ...(props.subscriptions?.[sub.name] ?? {}),
-          memorySize: sub.runtimeProps?.memorySize,
-          timeout: sub.runtimeProps?.timeout
-            ? Duration.seconds(computeDurationSeconds(sub.runtimeProps.timeout))
-            : undefined,
-
-          role: props.subscriptions?.[sub.name]?.role ?? role,
-        });
-
-        if (sub.subscriptions.length > 0) {
-          // configure a Rule to route all subscribed events to the eventHandler
-          new Rule(handler, "Rules", {
-            eventBus: this.bus,
-            eventPattern: {
-              // only events that originate
-              // TODO: this seems like it would break service-to-service?
-              source: [this.serviceName],
-              detailType: Array.from(
-                new Set(sub.subscriptions.map((sub) => sub.name))
-              ),
-            },
-            targets: [
-              new LambdaFunction(handler, {
-                deadLetterQueue: this.deadLetterQueue,
-                retryAttempts: sub.runtimeProps?.retryAttempts,
-              }),
-            ],
-          });
-        }
-        return [sub.name, handler];
-      })
-    ) as {
-      [handler in EventHandlerNames<Service>]: Function;
-    };
-
-    this.handlers.forEach((handler) => {
-      this.configurePublish(handler);
-      this.configureEventHandler(handler);
-    });
-
-    if (this.handlers.length > 0) {
-      this.grantPrincipal = new CompositePrincipal(
-        ...Array.from(
-          this.handlers.reduce<Set<aws_iam.IPrincipal>>((roles, handler) => {
-            if (handler.role) {
-              roles.add(handler.role);
-            }
-            return roles;
-          }, new Set())
-        )
-      );
-    } else {
-      this.grantPrincipal = role;
-    }
   }
 
   public configurePublish(func: Function) {
@@ -188,19 +49,6 @@ export class Events<Service> extends Construct implements IGrantable {
   @grant()
   public grantPublish(grantable: IGrantable) {
     this.bus.grantPutEventsTo(grantable);
-  }
-
-  private configureEventHandler(handler: Function) {
-    // allows the access to all of the operations on the injected service client
-    this.props.service.configureForServiceClient(handler);
-    this.handlers.map((handler) =>
-      this.props.service.configureForServiceClient(handler as Function)
-    );
-    // allow http access to the service client
-    this.props.api.configureInvokeHttpServiceApi(handler);
-    this.handlers.map((handler) =>
-      this.props.api.configureInvokeHttpServiceApi(handler as Function)
-    );
   }
 
   private readonly ENV_MAPPINGS = {