+++ packages/@eventual/aws-runtime/src/handlers/api/middleware.ts
@@ -1,6 +1,10 @@
 import { extendsError } from "@eventual/core/internal";
 import middy, { MiddlewareObj } from "@middy/core";
-import { APIGatewayProxyHandlerV2 } from "aws-lambda";
+import type {
+  APIGatewayProxyEventV2,
+  APIGatewayProxyHandlerV2,
+  APIGatewayProxyResultV2,
+} from "aws-lambda";
 import util from "util";
 
 /**
@@ -28,6 +32,11 @@ export const errorMiddleware: MiddlewareObj = {
 
 export function withErrorMiddleware<TEvent>(
   handler: APIGatewayProxyHandlerV2<TEvent>
-) {
+): middy.MiddyfiedHandler<
+  APIGatewayProxyEventV2,
+  APIGatewayProxyResultV2<TEvent>,
+  Error,
+  any
+> {
   return middy(handler).use(errorMiddleware);
 }