+++ packages/@eventual/aws-runtime/src/handlers/event-handler.ts
@@ -2,7 +2,7 @@ import "@eventual/injected/entry";
 
 import {
   createEventHandlerWorker,
-  GlobalEventHandlerProvider,
+  GlobalSubscriptionProvider,
 } from "@eventual/core-runtime";
 import type { EventBridgeEvent } from "aws-lambda";
 import { createEventClient, createServiceClient } from "../create.js";
@@ -12,7 +12,7 @@ export const processEvent = createEventHandlerWorker({
   serviceClient: createServiceClient({
     eventClient: createEventClient(),
   }),
-  eventHandlerProvider: new GlobalEventHandlerProvider(),
+  eventHandlerProvider: new GlobalSubscriptionProvider(),
 });
 
 export default async function (event: EventBridgeEvent<string, any>) {