+++ packages/@eventual/client/src/index.ts
@@ -1,4 +1,5 @@
-export * from "./http-eventual-client.js";
 export * from "./base-http-client.js";
+export * from "./http-eventual-client.js";
+export * from "./http-request.js";
 export * from "./request-handler/index.js";
 export * from "./service-client.js";