+++ packages/@eventual/core-runtime/src/handlers/api-handler.ts
@@ -36,6 +36,7 @@ export function createApiHandler({ serviceClient }: ApiHandlerDependencies) {
    * then handles the request.
    */
   return function processRequest(request: HttpRequest): Promise<HttpResponse> {
+    console.log("request", request);
     return serviceTypeScope(ServiceType.ApiHandler, async () => {
       try {
         const response = await router.handle(request);
@@ -83,110 +84,158 @@ function initRouter() {
     const shouldValidate = command.validate !== false;
 
     // RPC route takes a POST request and passes the parsed JSON body as input to the input
-    router.post(`/_rpc/${command.name}`, async (request) => {
-      if (command.passThrough) {
-        // if passthrough is enabled, just proxy the request-response to the handler
-        return command.handler(request);
-      }
-
-      let input = await request.tryJson();
-      if (command.input && shouldValidate) {
-        try {
-          input = command.input.parse(input);
-        } catch (err) {
-          console.error("Invalid input", err, input);
-          return new HttpResponse(JSON.stringify(err), {
-            status: 400,
-            statusText: "Invalid input",
-          });
+    router.post(
+      `/_rpc/${command.name}`,
+      withMiddleware(async (request, context) => {
+        if (command.passThrough) {
+          // if passthrough is enabled, just proxy the request-response to the handler
+          return command.handler(request, context);
         }
-      }
 
-      let output = await command.handler(input, {
-        headers: request.headers,
-      });
-      if (command.output && shouldValidate) {
-        try {
-          output = command.output.parse(output);
-        } catch (err) {
-          console.error("RPC output did not match schema", output, err);
-          return new HttpResponse(JSON.stringify(err), {
-            status: 500,
-            statusText: "RPC output did not match schema",
-          });
+        let input: any = await request.tryJson();
+        if (command.input && shouldValidate) {
+          try {
+            input = command.input.parse(input);
+          } catch (err) {
+            console.error("Invalid input", err, input);
+            return new HttpResponse(JSON.stringify(err), {
+              status: 400,
+              statusText: "Invalid input",
+            });
+          }
         }
-      }
-      return new HttpResponse(JSON.stringify(output, jsonReplacer), {
-        status: 200,
-      });
-    });
+
+        let output: any = await command.handler(input, context);
+        if (command.output && shouldValidate) {
+          try {
+            output = command.output.parse(output);
+          } catch (err) {
+            console.error("RPC output did not match schema", output, err);
+            return new HttpResponse(JSON.stringify(err), {
+              status: 500,
+              statusText: "RPC output did not match schema",
+            });
+          }
+        }
+        return new HttpResponse(JSON.stringify(output, jsonReplacer), {
+          status: 200,
+        });
+      })
+    );
 
     if (command.path) {
       const method = (command.method?.toLocaleLowerCase() ??
         "all") as keyof Router;
 
       // REST routes parse the request according to the command's path/method/params configuration
-      router[method](command.path, async (request: HttpRequest) => {
-        if (command.passThrough) {
-          // if passthrough is enabled, just proxy the request-response to the handler
-          return command.handler(request);
-        }
+      router[method](
+        command.path,
+        withMiddleware(async (request: HttpRequest, context) => {
+          if (command.passThrough) {
+            // if passthrough is enabled, just proxy the request-response to the handler
+            return command.handler(request, context);
+          }
 
-        // first, get the body as pure JSON - assume it's an object
-        const body = await request.tryJson();
-        let input: any = {
-          ...request.params,
-          ...(body && typeof body === "object" ? body : {}),
-        };
+          // first, get the body as pure JSON - assume it's an object
+          const body = await request.tryJson();
+          let input: any = {
+            ...request.params,
+            ...(body && typeof body === "object" ? body : {}),
+          };
 
-        // parse headers/params/queries/body into the RPC interface
-        if (command.params) {
-          Object.entries(command.params as Record<string, RestParamSpec>).map(
-            ([name, spec]) => {
-              input[name] = resolveInput(name, spec);
+          // parse headers/params/queries/body into the RPC interface
+          if (command.params) {
+            Object.entries(command.params as Record<string, RestParamSpec>).map(
+              ([name, spec]) => {
+                input[name] = resolveInput(name, spec);
+              }
+            );
+          }
+
+          if (command.input && shouldValidate) {
+            // validate the zod input schema if one is specified
+            input = command.input.parse(input);
+          }
+
+          // call the command RPC handler
+          let output: any = await command.handler(input, context);
+
+          if (command.output && shouldValidate) {
+            // validate the output of the command handler against the schema if it's defined
+            output = command.output.parse(output);
+          }
+
+          // TODO: support mapping RPC output back to HTTP properties such as Headers
+          // TODO: support alternative status code https://github.com/functionless/eventual/issues/276
+
+          return new HttpResponse(JSON.stringify(output, jsonReplacer), {
+            status: 200,
+            headers: {
+              "Content-Type": "application/json",
+            },
+          });
+
+          function resolveInput(name: string, spec: RestParamSpec): any {
+            if (spec === "body") {
+              return body?.[name];
+            } else if (spec === "query") {
+              return request.query?.[name];
+            } else if (spec === "header") {
+              return request.headers.get(name);
+            } else if (spec === "path") {
+              return request.params?.[name];
+            } else {
+              return resolveInput(spec.name ?? name, spec.in);
             }
-          );
-        }
+          }
+        })
+      );
+    }
 
-        if (command.input && shouldValidate) {
-          // validate the zod input schema if one is specified
-          input = command.input.parse(input);
-        }
+    /**
+     * Applies the chain of middleware callbacks to the request to build up
+     * context and pass it through the chain and finally to the handler.
+     *
+     * Each context can add to or completely replace the context. They can
+     * also break the chain at any time by returning a HttpResponse instead
+     * of calling `next`.
+     *
+     * @param handler
+     * @returns
+     */
+    function withMiddleware(
+      handler: (request: HttpRequest, context: any) => Promise<HttpResponse>
+    ) {
+      return async (request: HttpRequest): Promise<HttpResponse> => {
+        const chain = (command.middlewares ?? []).values();
 
-        // call the command RPC handler
-        let output = await command.handler(input, {
-          headers: request.headers,
-        });
+        return next(request, {});
 
-        if (command.output && shouldValidate) {
-          // validate the output of the command handler against the schema if it's defined
-          output = command.output.parse(output);
-        }
-
-        // TODO: support mapping RPC output back to HTTP properties such as Headers
-        // TODO: support alternative status code https://github.com/functionless/eventual/issues/276
-
-        return new HttpResponse(JSON.stringify(output, jsonReplacer), {
-          status: 200,
-          headers: {
-            "Content-Type": "application/json",
-          },
-        });
-
-        function resolveInput(name: string, spec: RestParamSpec): any {
-          if (spec === "body") {
-            return body?.[name];
-          } else if (spec === "query") {
-            return request.query?.[name];
-          } else if (spec === "header") {
-            return request.headers?.[name];
-          } else if (spec === "path") {
-            return request.params?.[name];
+        async function next(
+          request: HttpRequest,
+          context: any
+        ): Promise<HttpResponse> {
+          let consumed = false;
+          const middleware = chain.next();
+          if (middleware.done) {
+            return handler(request, context);
           } else {
-            return resolveInput(spec.name ?? name, spec.in);
+            return middleware.value({
+              request,
+              context,
+              next: async (context) => {
+                if (consumed) {
+                  consumed = true;
+                  throw new Error(
+                    `Middleware cannot call 'next' more than once`
+                  );
+                }
+                return next(request, context);
+              },
+            });
           }
         }
-      });
+      };
     }
   }