+++ packages/@eventual/core-runtime/src/handlers/event-handler.ts
@@ -1,6 +1,10 @@
 import { EventEnvelope, EventualServiceClient } from "@eventual/core";
-import { registerServiceClient, serviceTypeScope, ServiceType } from "@eventual/core/internal";
-import { EventHandlerProvider } from "../providers/event-handler-provider.js";
+import {
+  registerServiceClient,
+  ServiceType,
+  serviceTypeScope,
+} from "@eventual/core/internal";
+import { SubscriptionProvider } from "../providers/subscription-provider.js";
 
 /**
  * The dependencies of {@link createEventHandlerWorker}.
@@ -14,7 +18,7 @@ export interface EventHandlerDependencies {
   /**
    * Returns event handlers
    */
-  eventHandlerProvider: EventHandlerProvider;
+  eventHandlerProvider: SubscriptionProvider;
 }
 
 export interface EventHandlerWorker {
@@ -37,12 +41,12 @@ export function createEventHandlerWorker({
   }
 
   return async function (events) {
-    return await serviceTypeScope(ServiceType.EventHandler, async () => {
+    return await serviceTypeScope(ServiceType.Subscription, async () => {
       await Promise.allSettled(
         events.map((event) =>
           Promise.allSettled(
             eventHandlerProvider
-              .getEventHandlersForEvent(event.name)
+              .getSubscriptionsForEvent(event.name)
               .map((handler) => handler(event.event))
           )
         )