+++ packages/@eventual/core/src/index.ts
@@ -13,6 +13,7 @@ export * from "./schedule.js";
 export * from "./secret.js";
 export * from "./service-client.js";
 export * from "./signals.js";
+export * from "./subscription.js";
 export * from "./workflow-events.js";
 export * from "./workflow.js";
 export * as internal from "./internal/index.js";